Only trigger select2 to change if the watcher value is different from select2 val
Currently the the input event will be triggered twice. First when the actual select changes and then once again when the value prop is updated via v-model triggering the value watcher to trigger the change event again even though nothing in the select has changed. With this change the value watcher only triggers a change of the select if the prop value is different than the select2.val().